making progress on how my front end should look

+++ construt2ls/construct2lsapp/models.py
@@ -16,7 +16,8 @@ class Product(models.Model):
     product_name = models.CharField(max_length=50)
     category = models.CharField(max_length=50, default="")
     subcategory = models.CharField(max_length=50, default="")
-    price = models.IntegerField(default=0)
+    #price = models.IntegerField(default=0)
+    price = models.DecimalField(max_digits=10, decimal_places=2, default=0)
     desc = models.CharField(max_length=300)
 
     image = models.ImageField(upload_to='images/images', default="")
